+++ closet/src/main/resources/templates/top/top.html
@@ -39,6 +39,7 @@
     
       <div class="slideshow">
 	      <div class="slideshow_js">
+	      <!-- if文入れる　+${items[0].picture}}がNULLの時の処理 -->
 		      <img th:src="@{'/images/top/'+${items[0].picture}}" class="slide_pic">
 	      </div>
       </div>
